Prevent timesheet component render on Login
