+++ src/Components/Employee/EmployeesList/employee.resource.tsx
@@ -1,6 +1,5 @@
 const url = process.env.REACT_APP_URL;
-const token = localStorage.getItem('token');
-export const getAllEmployees = async () => {
+export const getAllEmployees = async (token) => {
   const options = {
     headers: {
       Authorization: `Bearer  ${token}`,
