+++ src/Components/Employee/Profile/employee-profile.module.css
@@ -36,7 +36,6 @@
 }
 .zoom{
     display: block;
-    margin-left: 10%;
     margin-right: auto;
 }
 .card {
