--- src/Components/Navigation/NavigationBar.module.css
@@ -1,18 +1,0 @@
-.navbar {
-    background-color: #054242 !important;
-    border: solid 0.5px white;
-}
-
-.sideNav {
-    background-color: #054242 !important;
-    width: 100%;
-}
-
-.sideNavClosed {
-    width: 0px;
-}
-
-.h1 {
-    font-size: 12;
-    color: red;
-}
