+++ src/Components/Timesheets/timesheet.module.scss
@@ -1,5 +1,4 @@
 .form{
- margin-left:10%;
  width: 100%;
 }
 .toast {
